renaming api routing files to be more descriptive
